Add license message.
This code is published under the MIT license.
It still contains traces of pre-free things, e.g. that MathUtil.cs has an
expiration date, and License.cs has (unused) routines for license signature
verification (and a cert from rumi-sound).  There's also a lot of cruft in
other parts of the codebase, to be cleaned up some time...!